requirements-windows.txt
django
djangorestframework
environs
requests
psycopg2-binary
pyinstaller